create Router setting && components , pages
