+++ src/components/Detail.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+
+// 여기는 웹툰 디테일
+const Detail = () => {
+  return (
+    <div>
+      네 맞아요! detail
+    </div>
+  )
+}
+
+export default Detail;
