+++ src/components/Main.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+// 여기가 메인화면이에요
+const Main = () => {
+  return (
+    <div>
+      네 맞아요!
+    </div>
+  )
+}
+
+export default Main;
