+++ src/pages/DetailPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Detail from '../components/Detail';
+
+
+const DetailPage = () => {
+  return (
+    <Detail />
+  )
+}
+
+export default DetailPage;
