+++ src/pages/MainPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Main from '../components/Main';
+
+// 페이지는 components 폴더에 있는 파일을 불러와서 단순히 보여준다.
+const MainPage = () => {
+  return (
+    <Main />
+  )
+}
+
+export default MainPage;
